+++ packages/config-registry-controller/src/config-registry-api-service/types.ts
@@ -48,6 +48,7 @@ const ChainConfigSchema = type({
   isFeatured: boolean(),
   isDeprecated: boolean(),
   isDeletable: boolean(),
+  isAutoEnabled: boolean(),
   priority: number(),
 });